[JENKINS-31162] Polish before RC
