+++ core/src/main/resources/jenkins/model/Messages.properties
@@ -71,10 +71,3 @@ ParameterizedJobMixIn.build_now=Build Now
 BlockedBecauseOfBuildInProgress.shortDescription=Build #{0} is already in progress{1}
 BlockedBecauseOfBuildInProgress.ETA=\ (ETA:{0})
 BuildDiscarderProperty.displayName=Discard Old Builds
-
-ItemCategory.Uncategorized.DisplayName=Uncategorized
-ItemCategory.Uncategorized.Description=Item types that have not yet been categorized by their plugin maintainer.
-ItemCategory.StandaloneProjects.DisplayName=Standalone Projects
-ItemCategory.StandaloneProjects.Description=Create projects with a self-contained configuration and history. These projects can be at the top-level or grouped within folders.
-ItemCategory.NestedProjects.DisplayName=Nested Projects
-ItemCategory.NestedProjects.Description=Create project categories or project hierarchies with folders. Folders can be created manually or automatically based on repositories.
